refactor: implement controller pattern across backend
