+++ backend/routes/skillsRoutes.js
@@ -1,56 +1,22 @@
 const express = require("express");
 const router = express.Router();
-const { UserSkill, Skill } = require("../models/schema");
-const { v4: uuidv4 } = require("uuid");
 const isAuthenticated = require("../middlewares/isAuthenticated");
 const authorizeRole = require("../middlewares/authorizeRole");
 const { ROLE_GROUPS } = require("../utils/roles");
+const skillController = require("../controllers/skillController");
+
 // GET unendorsed user skills for a particular skill type
 router.get(
   "/user-skills/unendorsed/:type",
   isAuthenticated,
-  async (req, res) => {
-    const skillType = req.params.type; // e.g. "cultural", "sports"
-
-    try {
-      const skills = await UserSkill.find({ is_endorsed: false })
-        .populate({
-          path: "skill_id",
-          match: { type: skillType },
-        })
-        .populate("user_id", "personal_info.name username user_id") // optionally fetch user info
-        .populate("position_id", "title");
-
-      // Filter out null populated skills (i.e., skill type didn't match)
-      const filtered = skills.filter((us) => us.skill_id !== null);
-
-      res.json(filtered);
-    } catch (err) {
-      console.error(err);
-      res.status(500).json({ message: "Error fetching unendorsed skills." });
-    }
-  },
+  skillController.getUnendorsedUserSkills
 );
 
 router.post(
   "/user-skills/endorse/:id",
   isAuthenticated,
   authorizeRole(ROLE_GROUPS.ADMIN),
-  async (req, res) => {
-    const skillId = req.params.id;
-    try {
-      const userSkill = await UserSkill.findById(skillId);
-      if (!userSkill) {
-        return res.status(404).json({ message: "User skill not found" });
-      }
-      userSkill.is_endorsed = true;
-      await userSkill.save();
-      res.json({ message: "User skill endorsed successfully" });
-    } catch (err) {
-      console.error(err);
-      res.status(500).json({ message: "Error endorsing user skill" });
-    }
-  },
+  skillController.endorseUserSkill
 );
 
 // REJECT (delete) a user skill
@@ -58,67 +24,18 @@ router.post(
   "/user-skills/reject/:id",
   isAuthenticated,
   authorizeRole(ROLE_GROUPS.ADMIN),
-  async (req, res) => {
-    const skillId = req.params.id;
-
-    try {
-      const deletedSkill = await UserSkill.findByIdAndDelete(skillId);
-
-      if (!deletedSkill) {
-        return res.status(404).json({
-          message: "User skill not found",
-        });
-      }
-
-      res.json({
-        message: "User skill rejected and deleted successfully",
-      });
-    } catch (err) {
-      console.error("Error rejecting user skill:", err);
-      res.status(500).json({
-        message: "Error rejecting user skill",
-      });
-    }
-  },
+  skillController.rejectUserSkill
 );
 
 // GET all unendorsed skills by type
-router.get("/unendorsed/:type", isAuthenticated, async (req, res) => {
-  const skillType = req.params.type;
-
-  try {
-    const skills = await Skill.find({ type: skillType, is_endorsed: false });
-    res.json(skills);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ message: "Error fetching unendorsed skills." });
-  }
-});
+router.get("/unendorsed/:type", isAuthenticated, skillController.getUnendorsedSkills);
 
 // POST endorse a skill
 router.post(
   "/endorse/:id",
   isAuthenticated,
   authorizeRole(ROLE_GROUPS.ADMIN),
-  async (req, res) => {
-    const skillId = req.params.id;
-
-    try {
-      const skill = await Skill.findById(skillId);
-
-      if (!skill) {
-        return res.status(404).json({ message: "Skill not found" });
-      }
-
-      skill.is_endorsed = true;
-      await skill.save();
-
-      res.json({ message: "Skill endorsed successfully", skill });
-    } catch (err) {
-      console.error(err);
-      res.status(500).json({ message: "Failed to endorse skill." });
-    }
-  },
+  skillController.endorseSkill
 );
 
 // REJECT (delete) a skill
@@ -126,131 +43,22 @@ router.post(
   "/reject/:id",
   isAuthenticated,
   authorizeRole(ROLE_GROUPS.ADMIN),
-  async (req, res) => {
-    const skillId = req.params.id;
-
-    try {
-      const deletedSkill = await Skill.findByIdAndDelete(skillId);
-
-      if (!deletedSkill) {
-        return res.status(404).json({
-          message: "Skill not found",
-        });
-      }
-
-      res.json({
-        message: "Skill rejected and deleted successfully",
-      });
-    } catch (err) {
-      console.error("Error rejecting skill:", err);
-      res.status(500).json({
-        message: "Failed to reject skill",
-      });
-    }
-  },
+  skillController.rejectSkill
 );
 
 //get all endorsed skills
-router.get("/get-skills", isAuthenticated, async (req, res) => {
-  try {
-    const skills = await Skill.find({ is_endorsed: true });
-    res.json(skills);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ message: "Failed to get endorsed skills." });
-  }
-});
+router.get("/get-skills", isAuthenticated, skillController.getAllSkills);
 
 //get all user skills (endorsed + unendorsed)
-router.get("/user-skills/:userId", isAuthenticated, async (req, res) => {
-  const userId = req.params.userId;
-  try {
-    const userSkills = await UserSkill.find({ user_id: userId })
-      .populate("skill_id")
-      .populate({
-        path: "position_id",
-        populate: {
-          path: "unit_id",
-          select: "name",
-        },
-      });
-    res.json(userSkills);
-  } catch (err) {
-    console.error("Failed to get user skills:", err);
-    res.status(500).json({ message: "Failed to get user skills." });
-  }
-});
+router.get("/user-skills/:userId", isAuthenticated, skillController.getUserSkills);
 
 //create a new skill
-router.post("/create-skill", isAuthenticated, async (req, res) => {
-  try {
-    const { name, category, type, description } = req.body;
-    const skill = new Skill({
-      skill_id: uuidv4(),
-      name,
-      category,
-      type,
-      description,
-    });
-    await skill.save();
-    res.status(201).json(skill);
-  } catch (err) {
-    res.status(500).json({ error: "Failed to add skill" });
-  }
-});
+router.post("/create-skill", isAuthenticated, skillController.createSkill);
 
 //create new user skill
-router.post("/create-user-skill", isAuthenticated, async (req, res) => {
-  try {
-    const { user_id, skill_id, proficiency_level, position_id } = req.body;
-
-    const newUserSkill = new UserSkill({
-      user_id,
-      skill_id,
-      proficiency_level,
-      position_id: position_id || null,
-    });
-
-    await newUserSkill.save();
-    res.status(201).json(newUserSkill);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: "Failed to add user skill" });
-  }
-});
+router.post("/create-user-skill", isAuthenticated, skillController.createUserSkill);
 
 // GET top 5 most popular skills campus-wide
-router.get("/top-skills", isAuthenticated, async (req, res) => {
-  try {
-    const topSkills = await UserSkill.aggregate([
-      { $match: { is_endorsed: true } },
-      { $group: { _id: "$skill_id", totalUsers: { $sum: 1 } } },
-      { $sort: { totalUsers: -1 } },
-      { $limit: 5 },
-      {
-        $lookup: {
-          from: "skills",
-          localField: "_id",
-          foreignField: "_id",
-          as: "skillDetails",
-        },
-      },
-      { $unwind: "$skillDetails" },
-      {
-        $project: {
-          _id: 0,
-          skillName: "$skillDetails.name",
-          type: "$skillDetails.type",
-          totalUsers: 1,
-        },
-      },
-    ]);
-
-    res.json(topSkills);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ message: "Error fetching top skills." });
-  }
-});
+router.get("/top-skills", isAuthenticated, skillController.getTopSkills);
 
 module.exports = router;
